Move Doctrine\Mapped base class to the root namespace DataSources
