+++ DataSources/Mapped.php
@@ -1,31 +1,21 @@
 <?php
 
-namespace DataGrid\DataSources\Doctrine;
-
-use Nette, Doctrine, DataGrid,
-	Doctrine\ORM\Query\Expr,
-	DataGrid\DataSources\IDataSource,
-	DataGrid\DataSources\DataSource;
+namespace DataGrid\DataSources;
 
 /**
  * Base class for Doctrine2 based data sources
- * 
  * @author Michael Moravec
  * @author Štěpán Svoboda
  */
 abstract class Mapped extends DataSource
 {
 	/**
-	 * Alias to column mapping
-	 *
-	 * @var array
+	 * @var array Alias to column mapping
 	 */
 	protected $mapping = array();
 
-
 	/**
 	 * Get columns mapping
-	 * 
 	 * @param array
 	 * @return void
 	 */
@@ -34,10 +24,8 @@ abstract class Mapped extends DataSource
 		return $this->mapping;
 	}
 
-	
 	/**
 	 * Set columns mapping
-	 *
 	 * @param array
 	 * @return void
 	 */
@@ -46,10 +34,8 @@ abstract class Mapped extends DataSource
 		$this->mapping = $mapping;
 	}
 
-
 	/**
 	 * Does datasource have column of given name?
-	 *
 	 * @return boolean
 	 */
 	public function hasColumn($name)
@@ -57,10 +43,8 @@ abstract class Mapped extends DataSource
 		return array_key_exists($name, $this->mapping);
 	}
 
-	
 	/**
 	 * Get list of column aliases
-	 *
 	 * @return array
 	 */
 	public function getColumns()
